RTX: Fix GCC ASM .file and .include for *.S
The GCC ASM files were renamed from *.s to *.S. However, the names by
which these files are referenced were not completely updated. This leads
to problems on case sensitive file systems.

Update the name by which the GCC ASM files are referenced, so as not to
break the build on case sensitive file systems.

Fixes: 7ae09c6c6911 "RTX: renamed GCC ASM files extension (*.s to *.S)"